+++ AI/static/client.js
@@ -1,0 +1,128 @@
+let messages = [];
+
+//Used to get locally stored chat with AI if any 
+const savedMessages = localStorage.getItem("chatMessages");
+if (savedMessages) {
+    messages = JSON.parse(savedMessages);
+    updateDisplayedMessages();
+}
+
+document.getElementById("messageInput").addEventListener("keypress", function (event) {
+    if (event.key === "Enter" && event.shiftKey) {
+
+    }
+    else if (event.key == "Enter") {
+        event.preventDefault(); 
+        sendAndRecieveMessageAi(); 
+    }
+});
+
+function sendAndRecieveMessageAi() {
+    const messageToSend = document.getElementById('messageInput').value;
+    if (messageToSend.trim() !== "") {
+        document.getElementById("messageInput").value = "";
+        messages.push({ user: messageToSend, response: null });
+        updateDisplayedMessages();
+        fetch('/post_message', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify({ message: messageToSend })
+        })
+            .then(response => response.json())
+            .then(data => {
+                messages[messages.length - 1].response = data.response;
+                updateDisplayedMessages();
+            })
+            .catch(error => console.log('Error:', error));
+    }
+}
+
+function updateDisplayedMessages() {
+    const displayMessagesElement = document.getElementById("displayMessages");
+    displayMessagesElement.innerHTML = "";
+    messages.forEach((msgObj) => {
+        const userMessageElement = document.createElement("p");
+        userMessageElement.innerText = msgObj.user;
+        displayMessagesElement.appendChild(userMessageElement);
+        if (msgObj.response) {
+            const responseMessageElement = document.createElement("p");
+            responseMessageElement.innerText = msgObj.response;
+            responseMessageElement.style.color = "#4a90e2"; 
+            displayMessagesElement.appendChild(responseMessageElement);
+        }
+    });
+    displayMessagesElement.scrollTop = displayMessagesElement.scrollHeight;
+    localStorage.setItem("chatMessages", JSON.stringify(messages));
+}
+
+
+//Main purpose of this is just to clear the local memory if the server is down when a user refreshes
+window.addEventListener('beforeunload', function (event) {
+    checkServerStatus();
+});
+
+document.addEventListener("DOMContentLoaded", function () {
+    if (performance.getEntriesByType("navigation")[0].type === "reload") {
+        console.log("The page was refreshed.");
+        checkServerStatus();
+    } else {
+        console.log("called resetChat");
+        resetChat();
+    }
+});
+
+function checkServerStatus() {
+    fetch('/status')
+        .then(response => {
+            if (!response.ok && response.status !== 304) {
+                throw new Error('Server is down');
+            }
+        })
+        .catch(error => {
+            console.log(error)
+            localStorage.clear();
+        });
+}
+
+function resetChatFromButton() {
+    let userresponse = confirm("This will reset your chat. Proceed?");
+    if (userresponse) {
+        resetChat();
+    }
+    else {
+        console.log("abort");
+    }
+}
+function resetChat() {
+    fetch('/reset_chat', {
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json'
+        }
+    })
+        .then(response => response.json())
+        .then(data => {
+            console.log(data.message); 
+            localStorage.clear();
+            location.reload()
+        })
+        .catch(error => {
+            console.error('Error:', error);
+        });
+}
+
+function submitUserCode() {
+    const bigMessageToSend = document.getElementById('bigMessageInput').value;
+    const url = '/submit_code_intermediary';  // This points to server.py, not directly to Docker
+    if (bigMessageToSend.trim() !== "") {
+        axios.post(url, { code: bigMessageToSend })  // Sending the C code to the Python intermediary server
+            .then(response => {
+                console.log(response.data);  // Print the response (test output or errors)
+            })
+            .catch(error => {
+                console.error('Error:', error.response ? error.response.data : error.message);
+            });
+    }
+}
