+++ AI/templates/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Flask-JS Communication</title>
+    <title>Prototype</title>
 </head>
 <body>
     <div class="main-container">
@@ -15,7 +15,7 @@
             <div class="input-container">
                 <textarea type="text" id="messageInput" placeholder="Enter a message"></textarea>
             <div class="sendandresbutton">
-                <button onclick="postMessage()">Send</button>
+                <button onclick="sendAndRecieveMessageAi()">Send</button>
                 <button class="secondbutton" onclick="resetChatFromButton()">Reset chat</button>
             </div>
             </div>
@@ -24,12 +24,12 @@
         <!-- Window 2: Big Text Input Window -->
         <div class="window">
             <h1>Code</h1>
-            <textarea id="bigMessageInput" placeholder="Put your code here once you are ready to submit"></textarea>
-            <button onclick="postBigMessage()">Submit</button>
+            <textarea id="bigMessageInput" placeholder="Write your code here"></textarea>
+            <button onclick="submitUserCode()">Submit</button>
         </div>
     </div>
 </body>
 <link rel="stylesheet" href="{{url_for('static', filename='style.css')}}">
-<script src="{{url_for('static', filename='something.js')}}"></script>
+<script src="{{url_for('static', filename='client.js')}}"></script>
 <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
 </html>
