--- Quiz/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Quiz Application</title>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <div class="quiz-container">
-        <div id="progress-bar">
-            <div id="progress"></div>
-        </div>
-        <div id="quiz">
-        </div>
-        <div id="ai-suggestion">
-            
-        </div>
-    </div>
-    <script src="scripts/script.js"></script>
-</body>
-</html>
