--- Quiz/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "compilerOptions": {
-      "target": "ES6",
-      "module": "ES6",
-      "lib": ["DOM", "ES6"],
-      "strict": true,
-      "esModuleInterop": true,
-      "skipLibCheck": true,
-      "outDir": "./dist"
-    },
-    "include": ["scripts/script.ts"],
-    "exclude": ["node_modules"]
-  }
-
